+++ lib/modules/settings/modules/resources.dart
@@ -2,6 +2,7 @@ import 'dart:io';
 import 'package:fluro/fluro.dart';
 import 'package:flutter/material.dart';
 import 'package:lunasea/core.dart';
+import 'package:lunasea/modules/settings.dart';
 
 class SettingsResourcesRouter extends LunaPageRouter {
     SettingsResourcesRouter() : super('/settings/resources');
@@ -19,58 +20,75 @@ class _State extends State<_SettingsResourcesRoute> {
     final GlobalKey<ScaffoldState> _scaffoldKey = GlobalKey<ScaffoldState>();
 
     @override
-    Widget build(BuildContext context) => Scaffold(
-        key: _scaffoldKey,
-        appBar: _appBar,
-        body: _body,
-    );
+    Widget build(BuildContext context) {
+        return Scaffold(
+            key: _scaffoldKey,
+            appBar: _appBar(),
+            body: _body(),
+        );
+    }
 
-    Widget get _appBar => LunaAppBar(title: 'Resources');
+    Widget _appBar() {
+        return LunaAppBar(
+            title: 'Resources',
+            state: context.read<SettingsState>(),
+        );
+    }
 
-    Widget get _body => LSListView(
-        children: [
-            LSCardTile(
-                title: LSTitle(text: 'Discord'),
-                subtitle: LSSubtitle(text: 'Chat & Discussions'),
-                trailing: LSIconButton(icon: CustomIcons.discord),
-                onTap: () async => await Constants.URL_DISCORD.lunaOpenGenericLink(),
-            ),
-            LSCardTile(
-                title: LSTitle(text: 'Documentation'),
-                subtitle: LSSubtitle(text: 'View the Documentation'),
-                trailing: LSIconButton(icon: CustomIcons.documentation),
-                onTap: () async => await Constants.URL_DOCUMENTATION.lunaOpenGenericLink(),
-            ),
-            LSCardTile(
-                title: LSTitle(text: 'Feedback Board'),
-                subtitle: LSSubtitle(text: 'Request New Features'),
-                trailing: LSIconButton(icon: Icons.speaker_notes),
-                onTap: () async => await Constants.URL_FEEDBACK.lunaOpenGenericLink(),
-            ),
-            LSCardTile(
-                title: LSTitle(text: 'GitHub'),
-                subtitle: LSSubtitle(text: 'View the Source Code'),
-                trailing: LSIconButton(icon: CustomIcons.github),
-                onTap: () async => await Constants.URL_GITHUB.lunaOpenGenericLink(),
-            ),
-            LSCardTile(
-                title: LSTitle(text: 'Reddit'),
-                subtitle: LSSubtitle(text: 'Ask Questions & Get Support'),
-                trailing: LSIconButton(icon: CustomIcons.reddit),
-                onTap: () async => await Constants.URL_REDDIT.lunaOpenGenericLink(),
-            ),
-            if(Platform.isIOS) LSCardTile(
-                title: LSTitle(text: 'TestFlight'),
-                subtitle: LSSubtitle(text: 'Join the TestFlight Beta'),
-                trailing: LSIconButton(icon: Icons.developer_board),
-                onTap: () async => await Constants.URL_TESTFLIGHT.lunaOpenGenericLink()
-            ),
-            LSCardTile(
-                title: LSTitle(text: 'Website'),
-                subtitle: LSSubtitle(text: 'Visit ${Constants.APPLICATION_NAME}\'s Website'),
-                trailing: LSIconButton(icon: CustomIcons.home),
-                onTap: () async => await Constants.URL_WEBSITE.lunaOpenGenericLink(),
-            ),
-        ],
-    );
+    Widget _body() {
+        return LunaListView(
+            scrollController: context.read<SettingsState>().scrollController,
+            children: [
+                LunaListTile(
+                    context: context,
+                    title: LunaText.title(text: 'Discord'),
+                    subtitle: LunaText.subtitle(text: 'Chat & Discussions'),
+                    trailing: LunaIconButton(icon: CustomIcons.discord),
+                    onTap: () async => await Constants.URL_DISCORD.lunaOpenGenericLink(),
+                ),
+                LunaListTile(
+                    context: context,
+                    title: LunaText.title(text: 'Documentation'),
+                    subtitle: LunaText.subtitle(text: 'View the Documentation'),
+                    trailing: LunaIconButton(icon: CustomIcons.documentation),
+                    onTap: () async => await Constants.URL_DOCUMENTATION.lunaOpenGenericLink(),
+                ),
+                LunaListTile(
+                    context: context,
+                    title: LunaText.title(text: 'Feedback Board'),
+                    subtitle: LunaText.subtitle(text: 'Request New Features'),
+                    trailing: LunaIconButton(icon: Icons.speaker_notes),
+                    onTap: () async => await Constants.URL_FEEDBACK.lunaOpenGenericLink(),
+                ),
+                LunaListTile(
+                    context: context,
+                    title: LunaText.title(text: 'GitHub'),
+                    subtitle: LunaText.subtitle(text: 'View the Source Code'),
+                    trailing: LunaIconButton(icon: CustomIcons.github),
+                    onTap: () async => await Constants.URL_GITHUB.lunaOpenGenericLink(),
+                ),
+                LunaListTile(
+                    context: context,
+                    title: LunaText.title(text: 'Reddit'),
+                    subtitle: LunaText.subtitle(text: 'Ask Questions & Get Support'),
+                    trailing: LunaIconButton(icon: CustomIcons.reddit),
+                    onTap: () async => await Constants.URL_REDDIT.lunaOpenGenericLink(),
+                ),
+                if(Platform.isIOS) LunaListTile(
+                    context: context,
+                    title: LunaText.title(text: 'TestFlight'),
+                    subtitle: LunaText.subtitle(text: 'Join the TestFlight Beta'),
+                    trailing: LunaIconButton(icon: Icons.developer_board),
+                    onTap: () async => await Constants.URL_TESTFLIGHT.lunaOpenGenericLink()
+                ),
+                LunaListTile(
+                    context: context,
+                    title: LunaText.title(text: 'Website'),
+                    subtitle: LunaText.subtitle(text: 'Visit ${Constants.APPLICATION_NAME}\'s Website'),
+                    trailing: LunaIconButton(icon: CustomIcons.home),
+                    onTap: () async => await Constants.URL_WEBSITE.lunaOpenGenericLink(),
+                ),
+            ],
+        );
+    }
 }